+++ openmodelica_microgrid_gym/env/__init__.py
@@ -1,1 +1,2 @@
 from .modelica import *
+from .recorder import *